B:C compatible monkey_patch implementation
We remove the call to set_subname and force it into runtime this permits
execution under the perl compiler. In order to do this we need to use a
closure and `goto CODEREF`. In order to do that, we must set the prototype
of the closure. That closure needs the same name to pass t/mojo/util.t --
this requires both calls to set_subname